Day 21: Monkey Math - Part 2

+++ src/day21/NumberMonkey.java
@@ -2,16 +2,15 @@ package day21;
 
 public class NumberMonkey extends AbstractMonkey {
 
-    private final long number;
+    private final double number;
 
     public NumberMonkey(String name, long number) {
         super(name);
         this.number = number;
     }
 
-
     @Override
-    public long getNumber() {
+    public double getNumber() {
         return number;
     }
 }
